requirements.txt
streamlit
vertexai
google-genai
google
time
os
trackback
python-dotenv